pyproject.toml
[project]
name = "oci-devday-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ipykernel>=6.29.5",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-google-genai>=2.1.3",
    "langgraph>=0.3.34",
    "oci>=2.150.2",
    "oracledb>=3.1.0",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "streamlit>=1.45.0",
    "traceloop-sdk>=0.39.4",
]
